separated Header and Footer testing
